Add multi-level lasers and boss level support

+++ Fight Sequence Version 0/src/types.h
@@ -56,7 +56,7 @@ struct Slime
     float height;
     float health;
     bool moveRight;
-    int speed;
+    float speed;
 };
 struct Spike
 {
